# Conflicts:
#	project_cw2/app/Http/Controllers/Admin/Users/LoginController.php
#	project_cw2/app/Http/Controllers/Customer/LoginCustomerController.php
#	project_cw2/app/Http/Controllers/ProductControllerView.php
#	project_cw2/app/Http/Services/CartService.php
#	project_cw2/app/Http/Services/Menu/MenuService.php
#	project_cw2/composer.lock
#	project_cw2/database/migrations/2021_05_29_085033_create_products_table.php
#	project_cw2/resources/views/carts/list.blade.php
#	project_cw2/resources/views/customer/history/profile.blade.php
#	project_cw2/resources/views/header.blade.php
#	project_cw2/resources/views/menu.blade.php
#	project_cw2/resources/views/products/content.blade.php
#	project_cw2/routes/web.php

+++ project_cw2/database/migrations/2021_05_29_085033_create_products_table.php
@@ -19,7 +19,7 @@ class CreateProductsTable extends Migration
             $table->text('description');
             $table->longText('content');
             $table->integer('menu_id');
-            $table->integer('price')->nullable();
+            $table->double('price')->nullable();
             $table->integer('price_sale')->nullable();
             $table->integer('active');
             $table->timestamps();
